refactor(analysis): externalize matrix3 inline affine tests
