+++ foundation/analysis/src/linalg/matrix/mod.rs
@@ -8,6 +8,9 @@ pub mod matrix2; // 2x2行列
 pub mod matrix3; // 3x3行列
 pub mod matrix4; // 4x4行列
 
+#[cfg(test)]
+pub mod matrix3_affine_tests;
+
 pub use matrix2::{Matrix2x2, Matrix2x2d, Matrix2x2f};
 pub use matrix3::{Matrix3x3, Matrix3x3d, Matrix3x3f};
 pub use matrix4::{Matrix4x4, Matrix4x4d, Matrix4x4f};
